Rename TxDetailsStorage to ScyllaDbTxDetailsStorage
