+++ tx-details-storage/src/drivers/mod.rs
@@ -1,0 +1,1 @@
+pub mod scylladb;
